requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
argcomplete==3.5.1
beautifulsoup4==4.12.3
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
distro==1.9.0
duckduckgo_search==7.1.0
fastapi==0.115.6
feedparser==6.0.11
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
googlesearch-python==1.2.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
packaging==24.1
pandas==2.2.3
pgvector==0.3.6
phidata==2.7.5
pillow==11.0.0
pipx==1.7.1
primp==0.9.2
psycopg2-binary==2.9.10
pycountry==24.6.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
rich==13.9.4
s3transfer==0.10.4
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.1
tweepy==4.14.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
google-generativeai